% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_import.R
\name{raw_data_import}
\alias{raw_data_import}
\title{Import raw data from github}
\usage{
raw_data_import()
}
\description{
Import raw data from github
}
\details{
This function is used to support investigation into
the data cleaning process. It should only be used after cloning the evaluationDB repository
at \url{https://github.com/williamlief/evaluationDB}.

It downloads the original government reports and FOIA request responses used
to create the evaluationDB
from \url{https://github.com/williamlief/evaluationDB-Raw}.

Data files are installed into the `data-raw/` directory and supplies all files
necessary for the data cleaning scripts in `data-raw/scripts`
}
